requirements.txt
anndata==0.9.2
biopython==1.79    # 安装和导入的名称不同，导入是Bio
h5py==3.1.0
matplotlib==3.5.1
numpy==1.19.2
pandas==1.4.1
pysam==0.9.1
rpy2==3.5.17
scanpy==1.9.5
scikit-learn==1.0.2  # 安装和导入的名称不同，导入是sklearn
scipy==1.8.0
setuptools==68.2.2
tensorflow==2.6.0
tqdm==4.66.1